Reorganize folder structure to follow fractal pattern.
