+++ src/pages/preferences/preferencesTabs/notificationTab/NotificationField.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Radio } from "antd";
 import styled from "@emotion/styled";
 import { Notifications } from "gql/generated/types";
-import { notificationFields } from "./fieldMaps";
+import { notificationFields } from "constants/fieldMaps";
 
 interface NotificationFieldProps {
   notification: string;
